p2

+++ 05.py
@@ -17,8 +17,8 @@ for line in sys.stdin:
                 break
 
 def merge(ri, rj):
-    u, v = ri.start, ri.start + ri.stop
-    x, y = rj.start, rj.start + rj.stop
+    u, v = ri.start, ri.stop
+    x, y = rj.start, rj.stop
     if v < x or y < u:
         return None
     else:
@@ -29,26 +29,37 @@ def merge(ri, rj):
 def key(r):
     return (r.start, r.stop)
 
-merged = {}
-for ri in ranges:
-    for rj in ranges:
-        ki, kj = key(ri), key(rj)
-        print(f"Considering {ki}, {kj}:", end=' ')
-        rm = merge(ri, rj)
-        if rm is None:
-            merged[key(ri)] = ri
-            print(f"no overlap, adding {ki}")
-        else:
-            km = key(rm)
-            print(f"overlap, adding {km}")
-            if key(ri) in merged:
-                del merged[key(ri)]
-                print(f"Removed {ki}")
-            if key(rj) in merged:
-                del merged[key(rj)]
-                print(f"Removed {kj}")
-            merged[key(rm)] = rm
+merged = None
+for _ in range(0, 5):
+    if merged is None:
+        merged = {}
+    else:
+        ranges = list(merged.values())
+    print(ranges)
+    for (i, ri) in enumerate(ranges):
+        for (j, rj) in enumerate(ranges):
+            if i == j: continue
+            ki, kj = key(ri), key(rj)
+            print(f"Considering {ki}, {kj}:", end=' ')
+            rm = merge(ri, rj)
+            if rm is None:
+                merged[key(ri)] = ri
+                print(f"no overlap, adding {ki}")
+            else:
+                km = key(rm)
+                print(f"overlap, adding {km}")
+                print(merged, key(ri))
+                if key(ri) in merged:
+                    del merged[key(ri)]
+                    print(f"Removed {ki}")
+                if key(rj) in merged:
+                    del merged[key(rj)]
+                    print(f"Removed {kj}")
+                merged[key(rm)] = rm
 
-# print(fc)
+print(fc)
 from pprint import pprint
+pprint(ranges)
 pprint(merged)
+
+print(sum([r.stop - r.start for r in ranges]))
